BUGS-6304 Update transaction counts after validation trim
When the validations run with the 'truncate' option on, the transaction
ST values are not updated with the correct number of segments. By
calling `prepare!` when the truncation option has been turned on, it
updates the particular segment (transaction, envelope, or otherwise)
with the new number.

Dev note: it appears that some of the test case data that is listed as
being valid is not valid, e.g. some ST01 values are either nonexistent
when they are required for the particular format, or they have
inaccurate values.